Fix type conversion warnings

+++ source/App/Parcat/parcat.cpp
@@ -379,7 +379,7 @@ std::vector<uint8_t> process_segment(const char * path, int idx, int * poc_base,
   }
 
   fseek(fdi, 0, SEEK_END);
-  int full_sz = ftell(fdi);
+  size_t full_sz = ftell(fdi);
   fseek(fdi, 0, SEEK_SET);
 
   std::vector<uint8_t> v(full_sz);
